+++ pkgs/record_use/lib/src/syntax.g.dart
@@ -281,59 +281,6 @@ extension CreationInstanceSyntaxExtension on InstanceSyntax {
       CreationInstanceSyntax.fromJson(json, path: path);
 }
 
-class DefinitionSyntax extends JsonObjectSyntax {
-  DefinitionSyntax.fromJson(
-    super.json, {
-    super.path = const [],
-  }) : super.fromJson();
-
-  DefinitionSyntax({
-    required IdentifierSyntax identifier,
-    String? loadingUnit,
-    super.path = const [],
-  }) : super() {
-    _identifier = identifier;
-    _loadingUnit = loadingUnit;
-    json.sortOnKey();
-  }
-
-  IdentifierSyntax get identifier {
-    final jsonValue = _reader.map$('identifier');
-    return IdentifierSyntax.fromJson(jsonValue, path: [...path, 'identifier']);
-  }
-
-  set _identifier(IdentifierSyntax value) {
-    json['identifier'] = value.json;
-  }
-
-  List<String> _validateIdentifier() {
-    final mapErrors = _reader.validate<Map<String, Object?>>('identifier');
-    if (mapErrors.isNotEmpty) {
-      return mapErrors;
-    }
-    return identifier.validate();
-  }
-
-  String? get loadingUnit => _reader.get<String?>('loading_unit');
-
-  set _loadingUnit(String? value) {
-    json.setOrRemove('loading_unit', value);
-  }
-
-  List<String> _validateLoadingUnit() =>
-      _reader.validate<String?>('loading_unit');
-
-  @override
-  List<String> validate() => [
-    ...super.validate(),
-    ..._validateIdentifier(),
-    ..._validateLoadingUnit(),
-  ];
-
-  @override
-  String toString() => 'DefinitionSyntax($json)';
-}
-
 class IdentifierSyntax extends JsonObjectSyntax {
   IdentifierSyntax.fromJson(
     super.json, {
@@ -874,12 +821,12 @@ class RecordingSyntax extends JsonObjectSyntax {
 
   RecordingSyntax({
     List<CallSyntax>? calls,
-    required DefinitionSyntax definition,
+    required IdentifierSyntax identifier,
     List<InstanceSyntax>? instances,
     super.path = const [],
   }) : super() {
     _calls = calls;
-    _definition = definition;
+    _identifier = identifier;
     _instances = instances;
     json.sortOnKey();
   }
@@ -918,21 +865,21 @@ class RecordingSyntax extends JsonObjectSyntax {
     return [for (final element in elements) ...element.validate()];
   }
 
-  DefinitionSyntax get definition {
-    final jsonValue = _reader.map$('definition');
-    return DefinitionSyntax.fromJson(jsonValue, path: [...path, 'definition']);
+  IdentifierSyntax get identifier {
+    final jsonValue = _reader.map$('identifier');
+    return IdentifierSyntax.fromJson(jsonValue, path: [...path, 'identifier']);
   }
 
-  set _definition(DefinitionSyntax value) {
-    json['definition'] = value.json;
+  set _identifier(IdentifierSyntax value) {
+    json['identifier'] = value.json;
   }
 
-  List<String> _validateDefinition() {
-    final mapErrors = _reader.validate<Map<String, Object?>>('definition');
+  List<String> _validateIdentifier() {
+    final mapErrors = _reader.validate<Map<String, Object?>>('identifier');
     if (mapErrors.isNotEmpty) {
       return mapErrors;
     }
-    return definition.validate();
+    return identifier.validate();
   }
 
   List<InstanceSyntax>? get instances {
@@ -973,7 +920,7 @@ class RecordingSyntax extends JsonObjectSyntax {
   List<String> validate() => [
     ...super.validate(),
     ..._validateCalls(),
-    ..._validateDefinition(),
+    ..._validateIdentifier(),
     ..._validateInstances(),
   ];